carent/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-extensions==3.1.5
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
